other updates after testing recent mac changes on the windows side

+++ src/cinder/audio2/SamplePlayer.cpp
@@ -86,7 +86,7 @@ double SamplePlayer::getLoopEndTime() const
 
 size_t SamplePlayer::getNumSeconds() const
 {
-	return (double)mNumFrames / (double)getSampleRate();
+	return size_t( (double)mNumFrames / (double)getSampleRate() );
 }
 
 // ----------------------------------------------------------------------------------------------------
